JAH - Carga de Datos finalizada - Historico Ventas

+++ 1.Layers/1.1Presentation/1.1.1Web/MVC.Presentacion/Models/Historico/HistoricoVentaModel.cs
@@ -10,7 +10,7 @@ namespace MVC.Presentacion.Models.Historico
         public int Id { get; set; }
         public int Mes { get; set; }
         public int Anio { get; set; }
-        public decimal MontoVenta { get; set; }
+        public double MontoVenta { get; set; }
         public bool EsPipa { get; set; }
         public bool EsCamioneta { get; set; }
         public bool EsLocal { get; set; }
